Update project structure and some filenames.
