+++ app/src/main/AndroidManifest.xml
@@ -10,9 +10,9 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".GraphicDesignActivity"
+            android:name=".ui.graphic_design.GraphicDesignActivity"
             android:label="@string/title_activity_graphic_design"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
