+++ app/src/main/java/com/example/webmasters/ui/MainActivity.java
@@ -1,6 +1,5 @@
-package com.example.webmasters;
+package com.example.webmasters.ui;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
@@ -8,6 +7,9 @@ import android.view.View;
 import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.webmasters.ui.graphic_design.GraphicDesignActivity;
+import com.example.webmasters.R;
 
 public class MainActivity extends AppCompatActivity {
