+++ app/src/main/java/com/example/webmasters/ui/graphic_design/DashboardFragment.java
@@ -1,4 +1,4 @@
-package com.example.webmasters.ui.home;
+package com.example.webmasters.ui.graphic_design;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,17 +14,14 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.webmasters.R;
 
-public class HomeFragment extends Fragment {
-
-    private HomeViewModel homeViewModel;
+public class DashboardFragment extends Fragment {
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        homeViewModel =
-                ViewModelProviders.of(this).get(HomeViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_home, container, false);
-        final TextView textView = root.findViewById(R.id.text_home);
-        homeViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+        DashboardViewModel dashboardViewModel = ViewModelProviders.of(this).get(DashboardViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_animations, container, false);
+        final TextView textView = root.findViewById(R.id.text_dashboard);
+        dashboardViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
