+++ app/src/main/java/com/example/webmasters/ui/graphic_design/GraphicDesignActivity.java
@@ -1,7 +1,8 @@
-package com.example.webmasters;
+package com.example.webmasters.ui.graphic_design;
 
 import android.os.Bundle;
 
+import com.example.webmasters.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -20,7 +21,7 @@ public class GraphicDesignActivity extends AppCompatActivity {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications)
+                R.id.navigation_logos, R.id.navigation_animations, R.id.navigation_themes)
                 .build();
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
