+++ app/src/main/res/layouts/graphic-design/layout/activity_graphic_design.xml
@@ -28,6 +28,6 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:navGraph="@navigation/mobile_navigation" />
+        app:navGraph="@navigation/graphic_design_navigation" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
