+++ app/src/main/res/navigation/graphic_design_navigation.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/mobile_navigation"
+    app:startDestination="@+id/navigation_logos">
+
+    <fragment
+        android:id="@+id/navigation_logos"
+        android:name="com.example.webmasters.ui.graphic_design.HomeFragment"
+        android:label="@string/title_home"
+        tools:layout="@layout/fragment_logos" />
+
+    <fragment
+        android:id="@+id/navigation_animations"
+        android:name="com.example.webmasters.ui.graphic_design.DashboardFragment"
+        android:label="@string/title_dashboard"
+        tools:layout="@layout/fragment_animations" />
+
+    <fragment
+        android:id="@+id/navigation_themes"
+        android:name="com.example.webmasters.ui.graphic_design.NotificationsFragment"
+        android:label="@string/title_notifications"
+        tools:layout="@layout/fragment_themes" />
+</navigation>
